requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
dnspython==2.2.1
Flask==2.1.1
fonttools==4.31.2
freezegun==1.2.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
neurokit2==0.1.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pymodm==0.4.3
pymongo==3.12.3
pyparsing==3.0.7
pytest==7.1.1
pytest-pycodestyle==2.3.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
urllib3==1.26.9
# windows-curses
Werkzeug==2.1.1
zipp==3.8.0
scikit-image